+++ packages/fe-fpm-writer/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @sap-ux/fe-fpm-writer
 
+## 1.3.2
+
+### Patch Changes
+
+#### Release Date
+
+2026-07-20
+
+#### Bug Fixes
+
+- Use onError instead of errorHandler in DOMParser options for xmldom 0.9.x compatibility [[7249031](https://github.com/SAP/open-ux-tools/commit/7249031c3c38bcc8c8afe1d448fd49268a3f4695)]
+
+## 1.3.1
+
+### Patch Changes
+
+#### Release Date
+
+2026-07-20
+
+#### Bug Fixes
+
+- Consolidate DOMParser instantiation behind getDOMParserOptions with correct errorHandler and xmlns support [[70ab106](https://github.com/SAP/open-ux-tools/commit/70ab106566883152b7483e96c297e5782a453e97)]
+
+## 1.3.0
+
+### Minor Changes
+
+#### Release Date
+
+2026-07-16
+
+#### Features
+
+- Add option to create Page Building Block with full layout during app generation [[85b88aa](https://github.com/SAP/open-ux-tools/commit/85b88aa64241d65d7a0618222221281bae2e3e45)]
+
 ## 1.2.1
 
 ### Patch Changes